chore(GRO-3003): decline 5-year searches for ClearScore
