+++ src/main/java/com/selina/lending/service/alternativeofferr/clearscore/ClearScoreAlternativeOfferRequestProcessor.java
@@ -4,7 +4,6 @@ import com.selina.lending.api.dto.common.LeadDto;
 import com.selina.lending.api.dto.qq.request.QuickQuoteApplicantDto;
 import com.selina.lending.service.alternativeofferr.AlternativeOfferRequestProcessor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
@@ -12,15 +11,11 @@ import java.util.List;
 
 @Slf4j
 @Component
-@ConditionalOnProperty(value = {
-        ClearScoreExtendedAlternativeOfferRequestProcessor.FEATURE_FLAG,
-        ClearScoreKeepOriginalRequestedLoanTermAlternativeOfferRequestProcessor.FEATURE_FLAG
-}, havingValue = "false", matchIfMissing = true)
-public class ClearScoreBasicAlternativeOfferRequestProcessor extends AlternativeOfferRequestProcessor {
+public class ClearScoreAlternativeOfferRequestProcessor extends AlternativeOfferRequestProcessor {
 
     private static final String CLIENT_ID = "clearscore";
-    private static final int MIN_ALTERNATIVE_OFFER_LOAN_TERM = 3;
-    private static final int MAX_ALTERNATIVE_OFFER_LOAN_TERM = 5;
+    private static final int ALTERNATIVE_OFFER_LOAN_TERM_5_YEARS = 5;
+    private static final int ALTERNATIVE_OFFER_LOAN_TERM_1_YEAR = 1;
 
     @PostConstruct
     private void init() {
@@ -39,13 +34,12 @@ public class ClearScoreBasicAlternativeOfferRequestProcessor extends Alternative
 
     @Override
     protected boolean isAlternativeRequestedLoanTerm(int requestedLoanTerm, List<QuickQuoteApplicantDto> applicants) {
-        return requestedLoanTerm >= MIN_ALTERNATIVE_OFFER_LOAN_TERM
-                && requestedLoanTerm < MAX_ALTERNATIVE_OFFER_LOAN_TERM;
+        return requestedLoanTerm == ALTERNATIVE_OFFER_LOAN_TERM_5_YEARS;
     }
 
     @Override
     protected int calculateAlternativeRequestedLoanTerm(int requestedLoanTerm, List<QuickQuoteApplicantDto> applicants) {
-        return MAX_ALTERNATIVE_OFFER_LOAN_TERM;
+        return ALTERNATIVE_OFFER_LOAN_TERM_1_YEAR;
     }
 
     @Override
